nsvfit updates + condor smoothing

+++ ZHtautauAnalysis/CONDOR/ShowStatus_condor.sh
@@ -5,7 +5,7 @@ outputDir=${3}
 time=`date "+%Y%m%d%H%M"`
 
 echo "killing jobs hanging for more than 2h:"
-source KillHangingJobs.sh 7200
+#source KillHangingJobs.sh 7200
 
 all=`ls ${name} | grep job | wc -l`
 echo "There are ${all} jobs in the task ${name}"
@@ -30,9 +30,12 @@ touch Resubmit_now.sh
 rm -f Resubmit_now.sh 
 echo "source Resubmit_condor.sh resubmit ${name}" > Resubmit_now.sh
 chmod +x Resubmit_now.sh
-echo "Removing jobs which are still running from the resubmission..."
-source ListRunningJobs.sh
-python FilterResubmit_task.py ${name}
-touch logFiles
-rm logFiles
-echo "To resubmit failed jobs do ./Resubmit_now.sh"
+n_resub=`more resubmit | wc -l`
+if [ $n_resub -gt 0 ]; then
+  echo "Removing jobs which are still running from the resubmission..."
+  source ListRunningJobs.sh
+  python FilterResubmit_task.py ${name}
+  touch logFiles
+  rm logFiles
+  echo "To resubmit failed jobs do ./Resubmit_now.sh"
+fi
